Add render file
