+++ src/render.js
@@ -1,0 +1,10 @@
+export const renderScores = (players) => {
+  const list = document.getElementById('scores-list');
+  players.forEach((player) => {
+    const { name, score } = player;
+    const li = document.createElement('li');
+    li.innerText = `${name} : ${score}`;
+    list.appendChild(li);
+  });
+  return list;
+};
